# service/src/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mage API
  description: Mage API
  contact:
    name: Mage Support
    email: mage@nga.mil
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: 6.0.1
servers: []
security:
  - mageToken: []
...
paths:
...
  /auth/{authenticationStrategy}/signin:
    post:
      tags: [ Authentication & Authorization ]
      summary: Authentication
      description: Authenticate a user using a given strategy.
      operationId: authenticate
      security:
        []
      parameters:
        - name: authenticationStrategy
          in: path
          description: Authentication strategy.
          required: true
          schema:
            type: string
            enum: [ local, ldap, saml, oauth, openidconnect ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserIn' }
      responses:
        200:
          description: Successfully signed in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Authentication' }
  /auth/token:
    post:
      tags: [ Authentication & Authorization  ]
      summary: Device Authorization
      description: Authorize a device.  If device does not exist, it may also be created via this method.
      operationId: authorize
      security:
        - authenticationToken: []
      parameters:
        - name: uid
          in: query
          description: Id of the device that requires provisioning
          required: true
          schema: { $ref: '#/components/schemas/Device/properties/id' }
        - name: name
          in: query
          description: Name of the device
          schema: { $ref: '#/components/schemas/Device/properties/name' }
        - name: description
          in: query
          description: Description of the device
          schema: { $ref: '#/components/schemas/Device/properties/description' }
        - name: appVersion
          in: query
          description: Application version of the device (required if creating a new device)
          schema: { $ref: '#/components/schemas/Device/properties/appVersion' }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/DeviceIn' }
      responses:
        200:
          description: Successful device authorization, Mage api token is returned.
        401:
          description:  User is not authorized.  Typically this means you need to authenticate `/auth/local/signin`.
        403:
          description: Failure to provision the device.  This typically means the device needs to be registered.
...
  /api/users:
    post:
      tags: [ User ]
      description: >
        Create a new user.  Duplicate usernames are not allowed.
        If the requesting user is an admin, i.e., has the `CREATE_USER`
        permission, the user record will be active.  Otherwise, the user will
        be inactive and an admin must activate the user later.  Additionally,
        the `roleId` key is required when the requesting user is an admin.
        Otherwise, the created user receives a default role.  When the request
        body contains `multipart/form-data`, Mage will handle the `icon` and/or
        `avatar` file form parameters.  The server only sets the user's map
        icon if the requesting user has the `CREATE_USER` permission.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        200:
          description: The created user document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    get:
      deprecated: true
      tags: [ User ]
      description: >
        This operation is deprecated and now only returns an empty response.  The client must have `READ_USER` permission.
      summary: Get an array of users.
      operationId: getUsers
      parameters:
        - in: query
          name: active
          description: Return only active or inactive users.
          schema:
            type: boolean
        - in: query
          name: enabled
          description: Return only enabled or disabled users.
          schema:
            type: boolean
        - in: query
          name: start
          description: Used with pagination.  Indicates the record number to start (leave blank for initial set of results).
          schema:
            type: integer
        - in: query
          name: limit
          description: Used with pagination.  Maximum number of records to return.
          schema:
            type: integer
        - in: query
          name: sort
          description: Describes the sort order.  Values allowed are asc, desc, ascending, descending, 1, and -1.
          content:
            application/json:
              schema:
                type: object
                required:
                  - displayName
                  - _id
                properties:
                  displayName:
                    type: string
                  _id:
                    type: string
        - in: query
          name: populate
          description: >
            Pre-populate the given relation keys in the result user documents.
            Currently, this only supports 'roleId'.
          schema: { $ref: '#/components/schemas/CommaSeparatedTokens' }
      responses:
        200:
          description: An array of user documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/users/count:
    get:
      tags: [ User ]
      description: >
        Return the number of users in the database.  The client must
        have `READ_USER` permission.
      summary: Count the number of users in the database.
      operationId: getUserCount
      parameters:
        - in: query
          name: active
          schema:
            type: boolean
        - in: query
          name: enabled
          schema:
            type: boolean
      responses:
        200:
          description: A successful request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Count' }
...
  /api/users/myself:
    get:
      tags: [ User ]
      description: >
        Return the user document for the requesting user based on the
        provided authentication token.
      operationId: getMyself
      responses:
        200:
          description: The user document for the requesting user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [ User ]
      description: >
        Update the user document for the requesting user based on the
        provided authentication token.  This operation does not support
        changing the requesting user's password; use
        `PUT /api/users/myself/password` for that.
      operationId: updateMyself
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateSelf' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserUpdateSelf' }
      responses:
        200:
          description: The updated user document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/myself/password:
    put:
      tags: [ User ]
      description: Update the password for the requesting user.
      operationId: updateMyPassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  format: password
                newPasswordConfirm:
                  type: string
                  format: password
      responses:
        200:
          description: Successful password update; return the user document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /api/users/myself/status:
    put:
      tags: [ User ]
      description: Update the status of the requesting user.
      operationId: updateMyStatus
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              required: [ status ]
      responses:
        200:
          description: Successful status update
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    delete:
      tags: [ User ]
      description: Delete the status of the requesting user.
      operationId: deleteMyStatus
      responses:
        200:
          description: Successfully deleted status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /api/users/{userId}:
    parameters:
      - { $ref: '#/components/parameters/userIdInPath' }
    get:
      tags:
      - User
      description: >
        Return the user document whose `id` equals the `userId` path parameter.
        The client must have `READ_USER` permission.
      operationId: getUserById
      responses:
        200:
          description: The updated user document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [ User ]
      description: >
        Update the user document whose `id` equals the `userId` path parameter.
        The client must have `UPDATE_USER` permission.
      operationId: updateUserById
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIn' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UserIn' }
      responses:
        200:
          description: user update response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [ User ]
      description: >
        Delete the user document whose `id` equals the `userId` path parameter.
        The client must have `DELETE_USER` permission.
      operationId: deleteUser
      responses:
        204:
          description: Success - user deleted
...
  /api/users/{userId}/events/{eventId}/recent:
    parameters:
      - { $ref: '#/components/parameters/userIdInPath' }
      - { $ref: '#/components/parameters/eventIdInPath' }
    post:
      tags: [ User ]
      description: >
        Add the given event to top of recent list for the given user.  The list
        is capped at 5.
      operationId: addRecentEventForUser
      responses:
        200:
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    authenticationToken:
      type: http
      description: >
        Token generated after successful authentication.  The client will exchange
        this token with an approved device for and API token.
      scheme: bearer
      bearerFormat: JWT
    mageToken:
      type: http
      description: >
        Mage generates an API token when a client authenticates
        initially.  The client can use the token for subsequent requests until
        it expires.
      scheme: bearer
  schemas:
    CommaSeparatedTokens:
      type: string
      pattern: '\w+(,\w+)*'
...
    Count:
      type: object
      required: [ count ]
      properties:
        count:
          type: integer
...
    UserIn:
      type: object
      properties:
        username: { $ref: '#/components/schemas/User/properties/username' }
        displayName: { $ref: '#/components/schemas/User/properties/displayName' }
        email: { $ref: '#/components/schemas/User/properties/email' }
        phone:
          type: string
          format: phone
        roleId: { $ref: '#/components/schemas/User/properties/roleId' }
        iconMetadata:
          type: object
          properties:
            type: { $ref: '#/components/schemas/UserIcon/properties/type' }
            color: { $ref: '#/components/schemas/UserIcon/properties/color' }
            text: { $ref: '#/components/schemas/UserIcon/properties/text' }
        icon:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
        avatar:
          description: >
            This key is valid only when encoded as `multipart/form-data`.
            This is a binary image file attachment.
          type: string
          format: binary
      allOf:
        - { $ref: '#/components/schemas/UserPasswordPair' }
    UserCreate:
      allOf:
        - { $ref: '#/components/schemas/UserIn' }
        - { type: object, required: [ username, displayName, password, passwordconfirm ] }
    UserUpdateSelf:
      allOf:
        - { $ref: '#/components/schemas/UserIn' }
        - type: object
          not:
            type: object
            required: [ password, passwordconfirm, icon, iconMetadata ]
...
    Authentication:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/UserRef'
        token:
          type: string
...
    DeviceIn:
      type: object
      required: [ uid ]
      properties:
        uid:
          type: string
        name:
          type: string
        description:
          type: string
        appVersion:
          type: string
        userId:
          $ref: '#/components/schemas/User/properties/id'
...
  parameters:
    userIdInPath:
      in: path
      name: userId
      description: The ID of the target user document
      required: true
      example: 5d0b2bfeeec24262f1a5fdf3
      schema: { $ref: '#/components/schemas/User/properties/id' }
    eventIdInPath:
      in: path
      name: eventId
      description: The ID of the target event document
      required: true
      example: 1234
      schema: { $ref: '#/components/schemas/Event/properties/id' }
...